% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.diagnostics_function.R
\name{mcmc.diagnostics}
\alias{mcmc.diagnostics}
\title{Markov Chain Monte Carlo Diagnostics}
\format{
See, function \code{nma.continuous.full.model}.
}
\usage{
mcmc.diagnostics(
  par,
  data,
  measure,
  assumption,
  mean.misspar,
  var.misspar,
  D,
  n.chains,
  n.iter,
  n.burnin,
  n.thin
)
}
\arguments{
\item{par}{A vector of the names of the parameters to be monitored. Up to three parameters can be named. See, section \strong{return} in function \code{nma.continuous.full.model}.}

\item{data}{A data-frame of a one-trial-per-row format containing arm-level data of each trial. This format is widely used for BUGS models. See 'Format' for the specification of the columns.}

\item{measure}{Character string indicating the effect measure with values \code{"MD"}, \code{"SMD"}, or \code{"ROM"}.}

\item{assumption}{Character string indicating the structure of the informative missingness parameter. Set \code{assumption} equal to one of the following: \code{"HIE-COMMON"}, \code{"HIE-TRIAL"}, \code{"HIE-ARM"}, \code{"IDE-COMMON"}, \code{"IDE-TRIAL"}, \code{"IDE-ARM"}, \code{"IND-CORR"}, or \code{"IND-UNCORR"}.}

\item{mean.misspar}{A positive non-zero number for the mean of the normal distribution of the informative missingness parameter.}

\item{var.misspar}{A positive non-zero number for the variance of the normal distribution of the informative missingness parameter.}

\item{D}{A binary number for the direction of the outcome. Set \code{direction = 0} for a positive outcome and \code{direction = 1} for a negative outcome.}

\item{n.chains}{Integer specifying the number of chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function.}

\item{n.iter}{Integer specifying the number of Markov chains for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function.}

\item{n.burnin}{Integer specifying the number of iterations to discard at the beginning of the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function.}

\item{n.thin}{Integer specifying the thinning rate for the MCMC sampling; an argument of the \code{\link[R2jags]{jags}} function.}
}
\value{
A panel of autocorrelation plots where the rows correspond to the chains and the columns correspond to the monitor parameters (maximum three).
Additionally, it uses the \code{\link[mcmcplots]{mcmcplot}} function to create an HTML file with a panel of diagnostic plots (trace, density, and autocorrelation) for each monitored parameter.
}
\description{
Markov Chain Monte Carlo Diagnostics
}
\examples{
### Obtain the diagnostic plots and check convergence for all monitored parameters using the R.hat
mcmc.diagnostics(par = c("tau2", "EM[3,1]", "EM[3,2]"), data = data, measure = "ROM", assumption = "HIE-TRIAL", mean.misspar = 0, var.misspar = 1, D = 0, n.chains = 3, n.iter = 10000, n.burnin = 1000, n.thin = 1)

}
\references{
Gelman, A, Rubin, DB. Inference from iterative simulation using multiple sequences. Stat Sci. 1992;7:457–472.

\dontshow{load("./data/One-stage model_NMA Dataset.RData")}
}
\seealso{
\code{\link{mcmcplots}}
}
